clean up for AccelStepper library
